+++ interface/Cargo.toml
@@ -1,0 +1,44 @@
+[package]
+name = "spl-token-interface"
+version = "1.0.0"
+description = "Solana Program Library Token Interface"
+documentation = "https://docs.rs/spl-token-interface"
+readme = "README.md"
+authors = { workspace = true }
+repository = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+arrayref = "0.3.9"
+bytemuck = { version = "1.23.1", features = ["derive"] }
+num-derive = "0.4"
+num_enum = "0.7.4"
+num-traits = { workspace = true }
+solana-instruction = "2.2.1"
+solana-program-error = "2.2.1"
+solana-program-option = "2.2.1"
+solana-program-pack = "2.2.1"
+solana-pubkey = "2.2.1"
+solana-sdk-ids = "2.2.1"
+spl-pod = { version = "0.5.1" }
+thiserror = "2.0"
+
+[dev-dependencies]
+proptest = "1.7"
+solana-native-token = "2.2.1"
+solana-pubkey = { version = "2.2.1", features = ["curve25519"] }
+strum = "0.26"
+strum_macros = "0.26"
+
+[lib]
+crate-type = ["lib"]
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[lints]
+workspace = true
+
+[package.metadata.solana]
+program-id = "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA"
