+++ interface/README.md
@@ -1,0 +1,14 @@
+# Token Interface
+
+A token program interface on the Solana blockchain, usable for fungible and
+non-fungible tokens.
+
+This crate provides an interface that third parties can utilize to create and
+use their tokens.
+
+Full documentation is available at [https://www.solana-program.com/docs/token](https://www.solana-program.com/docs/token)
+
+## Audit
+
+The audit repository [README](https://github.com/solana-labs/solana-program-library#audits)
+contains information about program audits.
